LPS-99992 Move upgrade to 3.0.2 so that it is executed on existing 7.2 installations. Register a dummy upgrade step for 1.1.8 to allow for backporting
