+++ modules/apps/journal/journal-service/src/main/java/com/liferay/journal/internal/upgrade/JournalServiceUpgrade.java
@@ -194,9 +194,7 @@ public class JournalServiceUpgrade implements UpgradeStepRegistrator {
 				_subscriptionLocalService, JournalArticle.class.getName(),
 				UpgradeDiscussionSubscriptionClassName.DeletionMode.ADD_NEW));
 
-		registry.register(
-			"1.1.7", "1.1.8",
-			new com.liferay.journal.internal.upgrade.v1_1_7.UpgradeUrlTitle());
+		registry.register("1.1.7", "1.1.8", new DummyUpgradeStep());
 
 		registry.register(
 			"1.1.8", "2.0.0",
@@ -215,6 +213,10 @@ public class JournalServiceUpgrade implements UpgradeStepRegistrator {
 			"3.0.0", "3.0.1",
 			new com.liferay.journal.internal.upgrade.v3_0_1.
 				UpgradeJournalArticle());
+
+		registry.register(
+			"3.0.1", "3.0.2",
+			new com.liferay.journal.internal.upgrade.v3_0_2.UpgradeUrlTitle());
 	}
 
 	protected void deleteTempImages() throws Exception {
